pyproject.toml
[tool.poetry]
name = "tad"
version = "0.1.0"
description = ""
authors = ["robbertuittenbroek <robbert.uittenbroek@rijksoverheid.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
questionary = "^2.0.1"
pyyaml = "^6.0.1"
jsonschema = "^4.21.1"
joblib = "^1.3.2"
scikit-learn = "^1.4.1.post1"
pandas = "^2.2.1"
shap = "^0.45.0"
matplotlib = "^3.8.3"
jinja2 = "^3.1.3"
pyyaml-include = "^2.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.3"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120


[tool.ruff.lint]
select = ["I", "SIM", "B", "UP", "F", "E"]
ignore = [] # List any rules to be ignored, currently empty.

# Allow auto-fixing of all enabled rules when using the `--fix` option.
fixable = ["ALL"]
unfixable = [] # Specify rules that cannot be auto-fixed, if any.

# Define a regex pattern for allowed unused variables (typically underscore-prefixed).
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.poetry.scripts]
pre-commit = 'pre-commit run --all-files:tad'

[tool.ruff.format]
# Enforce double quotes for strings, following Black's style.
quote-style = "double"

# Use spaces for indentation, in line with Black's formatting style.
indent-style = "space"

# Keep magic trailing commas, a feature of Black's formatting.
skip-magic-trailing-comma = false

# Automatically detect and use the appropriate line ending style.
line-ending = "auto"
